More quiz refactoring
