+++ quiz/create_quiz.py
@@ -3,7 +3,6 @@ import time
 import datetime
 import logging
 import threading
-import socket
 
 from common.config import initialize_config
 from quiz.quiz_queries import queries, connect_and_execute, insert_sqls, \
@@ -14,64 +13,6 @@ from quiz.type_map import get_type_from_id
 from providers.myssql_db import MySqlDB
 
 logger = logging.getLogger(__name__)
-
-
-def get_quiz_form_db(json_data):
-    user_id = json_data.get('user_id')
-    limit = json_data.get('limit', 1000)
-
-    # get all items by user
-    sql = queries[12].format(user_id, limit)
-    items = connect_and_execute(sql)
-    for item in items:
-        item['choices'] = json.loads(item.get('choices', {}))
-        item['metadata'] = json.loads(item.get('metadata', {}))
-        try:
-            item['topic'] = json.loads(item['topic'])
-            item['sub_topics'] = json.loads(item['sub_topics'])
-        except:
-            pass
-    # get all quizzes by user
-    sql = queries[13].format(user_id, limit)
-    exams = connect_and_execute(sql)
-    for item in exams:
-        item['metadata'] = json.loads(item.get('metadata', {}))
-
-    return {"items": items, "exams": exams,
-            "items_count": len(items), "exams_count": len(exams)}
-
-
-# returns items from DB for specific filters
-def get_items_db(json_data):
-    subject = json_data.get('subject')
-    topic = json_data.get('topic')
-    limit = json_data.get('limit', 50)
-    # keyword = json_data.get('keyword')
-    user_id = json_data.get('user_id', "")
-
-    sql = queries[9].format(subject, limit)
-    if len(user_id) > 2:
-        sql = queries[7].format(subject, limit, user_id)
-    if topic:
-        sql = queries[10].format(subject, topic, limit)
-        if len(user_id) > 2:
-            sql = queries[8].format(subject, topic, limit, user_id)
-
-    # print(sql)
-    results = connect_and_execute(sql)
-    results = sorted(results, key=lambda pos: pos['id'])
-
-    for result in results:
-        result['choices'] = json.loads(result.get('choices', {}))
-        result['metadata'] = json.loads(result.get('metadata', {}))
-        try:
-            result['topic'] = json.loads(result['topic'])
-            result['sub_topics'] = json.loads(result['sub_topics'])
-        except:
-            pass
-        result['answer'] = json.loads(result.get('answer', {}))
-
-    return {'items': results, 'total_items': len(results)}
 
 
 # extract and process items for quiz
@@ -88,8 +29,9 @@ def process_items(results):
         desc = metadata.get('quiz', '') + " " + topic
         try:
             topic = json.loads(topic)
-            topic.items()
-            tags.update(topic.items())
+            for t in topic.items():
+                if t[1]:
+                    tags.add(t[1])
             desc = metadata.get('description') or metadata.get('quiz', '')
         except:
             tags.add(topic)
@@ -106,7 +48,7 @@ def process_items(results):
         # print(item)
         questions.append(item)
         index += 1
-    return questions, tags
+    return questions, list(tags)
 
 
 def create_quiz_thread(title, desc, results, options, user_profile, new_id):
@@ -119,24 +61,26 @@ def create_quiz_thread(title, desc, results, options, user_profile, new_id):
     st = time.monotonic()
     try:
         items, tags = process_items(results)
-        tags = ','.join(tags)
-        # print(tags)
+        tags = str(','.join(tags))
+        #print(tags)
         email_name = options.get('required', 0x02)
+        searchable = options.get('searchable', searchable)
         options = {'email': email_name & 0x01, 'name': email_name & 0x02,
                    'show_correct': options.get('show_correct', 0)
                    }
-        searchable =  options.get('searchable', searchable)
+
         credentials = GoogleCredentials().get_credential()
         params = [title, desc, user_profile, items, options]
-        # print(params)
-        script_results = run_app_script(credentials, function_name='createQuiz',
-                                 params=params)
+        script_results = run_app_script(credentials,
+                                        function_name='createQuiz',
+                                        params=params)
         metadata = script_results.get('metadata', {})
         folder = metadata.get('folder_name')
         form_id = metadata.get('id')
         if form_id and folder:
             params = [form_id, folder]
-            run_app_script(credentials, function_name='moveFiles',
+            run_app_script(credentials,
+                           function_name='moveFiles',
                            params=params)
     except Exception as exc:
         logger.error("Quiz creation App Script exception: " + str(exc))
@@ -159,7 +103,7 @@ def create_quiz_thread(title, desc, results, options, user_profile, new_id):
                   json.dumps(items, indent=4),
                   metadata_quiz.get('published_url'),
                   json.dumps(user_profile, indent=4),
-                  user_profile.get('email'), tags, searchable
+                  user_profile.get('email'), tags, str(searchable)
                   )
         # print("****", values)
         db = MySqlDB()
@@ -204,31 +148,6 @@ def create_quiz_form_db(json_data, sql=None):
     except Exception as exc:
         logger.error("Quiz creation DB exception: " + str(exc))
         return {"quiz": {}, 'error': str(exc)}
-    # return {"quiz": results, "error": error, "elapsed": str(et)}
-
-
-def search_quiz(json_data):
-    id = json_data.get('id')
-    name = json_data.get('name')
-    keyword = json_data.get('keyword')
-    results = []
-    if id and name:
-        sql = queries[14].format(id, name)
-        results = connect_and_execute(sql)
-    if keyword:
-        sql = queries[15].format(keyword)
-        results = connect_and_execute(sql)
-
-    #print(len(results))
-
-    for result in results:
-        result['user_profile'] = json.loads(result.get('user_profile') or '{}')
-        result['metadata'] = json.loads(result.get('metadata') or '{}')
-
-    if len(results) > 0:
-        return {"quiz": results, "quiz_count": len(results)}
-    else:
-        return {"quiz": "Not found!", "quiz_count": 0}
 
 
 # sample quiz
@@ -288,10 +207,11 @@ if __name__ == '__main__':
     # http://api2.reliabilitymeasures.com/create_form/?input={"quiz_description":"Test","quiz_name":"Form 1","item_ids":[2,45,6,9,25]}
     # http://api2.reliabilitymeasures.com/get_items/?input={"subject":"Islam","limit":10}
 
-    ids = [1, 3, 6, 9, 25]
+    ids = [178,      179,      181,      182]
+
     # ids = [131,132]
     json_data = {'quiz_description': 'Test main', 'quiz_name': 'Form main',
-                 'item_ids': ids, 'options':{}}
+                 'item_ids': ids, 'options': {}}
     #print(json.dumps(create_quiz_form_db(json_data), indent=4))
 
     # json_data = {'user_id': "farrukh503@gmail.com", 'limit': 10}
@@ -307,11 +227,6 @@ if __name__ == '__main__':
     # print(json.dumps(res, indent=4))
 
     #print(create_quiz_sample({}))
-
-    json_data = {'id': 98, 'name': 'test'}
-    json_data = {'keyword': '2020'}
-    print(json.dumps(search_quiz(json_data), indent=4,
-                     default=decimal_default))
 
     et = time.monotonic() - st
     print(et)
